#249 : Added tests for reviews RUD view
